refactor(ContributionMonth): nest Monthly* maps under contribution-month, delete .old unused files
