+++ public/models/contribution-month/monthly-os-project.js
@@ -1,7 +1,7 @@
 import DefineMap from "can-define/map/";
 import DefineList from "can-define/list/";
-import OSProject from "./os-project";
-import ContributionMonth from "./contribution-month/";
+import OSProject from "../os-project";
+import ContributionMonth from "./contribution-month";
 import Observation from "can-observation";
 
 var MonthlyOSProject = DefineMap.extend("MonthlyOSProject",{
