+++ public/models/contribution-month/test/contribution-month_test.js
@@ -4,11 +4,11 @@ import moment from 'moment';
 
 import OSProject from '../../os-project';
 import ClientProject from '../../client-project';
-import MonthlyClientProjectOSProject from '../../monthly-client-projects-os-project';
+import MonthlyClientProjectOSProject from '../monthly-client-projects-os-project';
 import Contributor from '../../contributor';
-import MonthlyContributions from '../../monthly-contributions';
-import MonthlyOSProject from '../../monthly-os-project';
-import MonthlyClientProject from '../../monthly-client-project';
+import MonthlyContributions from '../monthly-contributions';
+import MonthlyOSProject from '../monthly-os-project';
+import MonthlyClientProject from '../monthly-client-project';
 
 QUnit.module( 'models/contribution-month' );
