+++ public/test/test.js
@@ -8,8 +8,8 @@ import 'bitcentive/components/contribution-month/contribution-month_test';
 import 'bitcentive/components/select-contribution-month/select-contribution-month_test';
 
 import 'bitcentive/components/login/login_test';
-//
+
 import 'bitcentive/components/register/register_test';
 
 import 'bitcentive/models/contribution-month/test/contribution-month_test';
-import 'bitcentive/models/monthly-os-project_test';
+import 'bitcentive/models/contribution-month/test/monthly-os-project_test';
